Unfitness calculation implemented

+++ Customer.cpp
@@ -10,17 +10,17 @@ struct Customer {
     double LAT;
     double LON;
     double angleWithDepot = 0.0;
-    double distanceFromDepot = 0;
+    uint32_t distanceFromDepot = 0;
     uint32_t demand = 0;
     bool isDepot = false;
-    static std::vector<double> depotDistances;
+    static std::vector<uint32_t> depotDistances;
     static std::vector<double> depotAngles;
     static std::vector<uint32_t> customerDemands;
     static std::vector<double> customerLAT;
     static std::vector<double> customerLON;
     static std::vector<uint32_t> customerIDs;
 };
-std::vector<double> Customer::depotDistances;
+std::vector<uint32_t> Customer::depotDistances;
 std::vector<double> Customer::depotAngles;
 std::vector<uint32_t> Customer::customerDemands;
 std::vector<double> Customer::customerLAT;
@@ -28,13 +28,13 @@ std::vector<double> Customer::customerLON;
 std::vector<uint32_t> Customer::customerIDs;
 
 
-std::vector<uint32_t> CIDsByDepotDistances(std::vector<double> depotDistances)
+std::vector<uint32_t> CIDsByDepotDistances(std::vector<uint32_t> depotDistances)
 {
-    std::multimap<uint32_t, double> mm;
+    std::multimap<uint32_t, uint32_t> mm;
     for (std::uint32_t i = 0; i != depotDistances.size(); ++i)
         mm.insert({ depotDistances[i], i });
 
-    std::vector<double> sortedAngles;
+    std::vector<uint32_t> sortedAngles;
     std::vector<std::uint32_t> indices;
     for (const auto& kv : mm) {
         sortedAngles.push_back(kv.first);
